+++ components/attention/w-attention.vue
@@ -1,5 +1,5 @@
 <script setup>
-import { watch, watchEffect, computed, ref, onMounted, nextTick, onUnmounted } from 'vue';
+import { watch, computed, ref, onMounted, nextTick, onUnmounted } from 'vue';
 
 import { i18n } from '@lingui/core';
 import { opposites, directions, autoUpdatePosition, useRecompute as recompute } from '@warp-ds/core/attention';
@@ -10,6 +10,7 @@ import { createModel, modelProps } from 'create-v-model';
 import { activateI18n } from '../util/i18n';
 
 import { props as attentionProps, getVariantClasses } from './attentionUtil.js';
+import { messages as daMessages } from './locales/da/messages.mjs';
 import { messages as enMessages } from './locales/en/messages.mjs';
 import { messages as fiMessages } from './locales/fi/messages.mjs';
 import { messages as nbMessages } from './locales/nb/messages.mjs';
@@ -17,7 +18,7 @@ import wAttentionArrow from './w-attention-arrow.vue';
 
 import { absentProp } from '#util';
 
-activateI18n(enMessages, nbMessages, fiMessages);
+activateI18n(enMessages, nbMessages, fiMessages, daMessages);
 
 defineOptions({
   name: 'wAttention',
@@ -62,10 +63,7 @@ const props = defineProps({
 });
 
 const emit = defineEmits(['update:modelValue', 'dismiss']);
-const attentionClasses = computed(() => ({
-  [props.attentionClass]: true,
-  [ccAttention.notCallout]: !props.callout,
-}));
+const attentionClasses = computed(() => [props.attentionClass, !props.callout && ccAttention.notCallout]);
 
 const wrapperClasses = computed(() => [ccAttention.base, getVariantClasses(props).wrapper]);
 
@@ -73,6 +71,7 @@ const model = props.modelValue === absentProp ? ref(true) : createModel({ props,
 const attentionEl = ref(null);
 const arrowEl = ref(null);
 const actualDirection = ref(props.placement);
+const targetElRef = ref(props.targetEl || null);
 
 const attentionState = computed(() => ({
   get isShowing() {
@@ -88,7 +87,7 @@ const attentionState = computed(() => ({
   directionName: props.placement,
   arrowEl: arrowEl.value?.$el,
   attentionEl: attentionEl.value,
-  targetEl: props.targetEl,
+  targetEl: targetElRef.value,
   noArrow: props.noArrow,
   distance: props.distance,
   skidding: props.skidding,
@@ -183,12 +182,27 @@ const defaultAriaLabel = computed(() => {
 
 let cleanup;
 
-onMounted(async () => {
-  watchEffect(model, recompute(attentionState.value), { immediate: props.callout });
+onMounted(() => {
+  // This watcher will only run in the client environment
+  // props.targetEl can be undefined if props.callout is true.
+  // However, the autoUpdatePosition() from @warp-ds/core, uses Floating-ui's computePosition(). Floating-ui's computePosition() requires a defined targetEl to be able to compute the attentionEl's position and the attentionEl's arrow position.
+  // When props.callout is true, we only need computePosition() to calculate the callout's arrow position. So, we create a default targetEl for callout that we can pass to the autoUpdatePosition(), in order to avoid Floating-ui from throwing an error.
+  watch(
+    () => [props.callout, props.targetEl],
+    ([callout, target]) => {
+      if (callout && target === undefined) {
+        targetElRef.value = document.createElement('div');
+      } else {
+        targetElRef.value = props.targetEl;
+      }
+    },
+    { immediate: true },
+  );
+  recompute(attentionState.value);
 });
 
 watch(
-  () => [props.targetEl, model.value, attentionEl.value],
+  () => [targetElRef.value, model.value, attentionEl.value],
   ([target, m, att]) => {
     if (!cleanup && m && target && att) {
       cleanup = autoUpdatePosition(attentionState.value);