+++ components/forms/w-select.vue
@@ -13,24 +13,21 @@ const p = defineProps(fieldProps);
 const emit = defineEmits(['update:modelValue']);
 const model = createModel({ props: p, emit });
 
-const wrapperClasses = computed(() => ({
-  [ccSelect.wrapper]: true,
-}));
+const chevronClasses = computed(() => [ccSelect.chevron, p.disabled && ccSelect.chevronDisabled]);
 
-const selectWrapperClasses = computed(() => ({
-  [ccSelect.selectWrapper]: true,
-}));
+const getSelectClasses = (hasValidationErrors) => [
+  ccSelect.base,
+  !hasValidationErrors && !p.disabled && !p.readOnly && ccSelect.default,
+  p.disabled && ccSelect.disabled,
+  hasValidationErrors && ccSelect.invalid,
+  p.readOnly && ccSelect.readOnly,
+];
 
-const selectClasses = computed(() => ({
-  [ccSelect.default]: true,
-  [ccSelect.disabled]: p.disabled,
-  [ccSelect.readOnly]: p.readOnly,
-}));
-
-const chevronClasses = computed(() => ({
-  [ccSelect.chevron]: true,
-  [ccSelect.chevronDisabled]: p.disabled,
-}));
+const handleKeyDown = (event) => {
+  if (p.readOnly && (event.key === ' ' || event.key === 'ArrowDown' || event.key === 'ArrowUp')) {
+    event.preventDefault();
+  }
+};
 </script>
 
 <script>
@@ -39,20 +36,15 @@ export default { name: 'wSelect', inheritAttrs: false };
 
 <template>
   <w-field v-slot="{ triggerValidation, aria, hasValidationErrors }" v-bind="{ ...$attrs, ...$props }">
-    <div :class="wrapperClasses">
-      <div :class="selectWrapperClasses">
+    <div :class="ccSelect.wrapper">
+      <div :class="ccSelect.selectWrapper">
         <select
           v-bind="{ ...aria, ...$attrs, class: '' }"
           :id="id"
           v-model="model"
-          :class="[
-            selectClasses,
-            {
-              [ccSelect.invalid]: hasValidationErrors,
-            },
-          ]"
+          :class="getSelectClasses(hasValidationErrors)"
           :disabled="disabled"
-          :readOnly="readOnly"
+          @keydown="handleKeyDown"
           @blur="triggerValidation">
           <slot />
         </select>