+++ components/util/i18n.ts
@@ -1,6 +1,6 @@
 import { Messages, i18n } from '@lingui/core';
 
-export const supportedLocales = ['en', 'nb', 'fi'] as const;
+export const supportedLocales = ['en', 'nb', 'fi', 'da'] as const;
 type SupportedLocale = (typeof supportedLocales)[number];
 
 export const defaultLocale = 'en';
@@ -20,7 +20,7 @@ function detectLocale(): SupportedLocale {
      * Server locale detection. This requires e.g LANG environment variable to be set on the server.
      */
     const serverLocale =
-      process.env.NMP_LANGUAGE ||
+      process?.env?.NMP_LANGUAGE ||
       Intl.DateTimeFormat().resolvedOptions().locale;
     return getSupportedLocale(serverLocale);
   }
@@ -38,10 +38,12 @@ export const getMessages = (
   locale: SupportedLocale,
   enMsg: Messages,
   nbMsg: Messages,
-  fiMsg: Messages
+  fiMsg: Messages,
+  daMsg: Messages
 ) => {
   if (locale === 'nb') return nbMsg;
   if (locale === 'fi') return fiMsg;
+  if (locale === 'da') return daMsg;
   // Default to English
   return enMsg;
 };
@@ -49,10 +51,11 @@ export const getMessages = (
 export const activateI18n = (
   enMessages: Messages,
   nbMessages: Messages,
-  fiMessages: Messages
+  fiMessages: Messages,
+  daMessages: Messages
 ) => {
   const locale = detectLocale();
-  const messages = getMessages(locale, enMessages, nbMessages, fiMessages);
+  const messages = getMessages(locale, enMessages, nbMessages, fiMessages, daMessages);
   i18n.load(locale, messages);
   i18n.activate(locale);
 };