+++ vite.config.js
@@ -1,3 +1,4 @@
+import rollupPluginReplace from '@rollup/plugin-replace';
 import vue from '@vitejs/plugin-vue';
 import { classes } from '@warp-ds/css/component-classes/classes';
 import { presetWarp } from '@warp-ds/uno';
@@ -14,14 +15,20 @@ export default defineConfig((env) => ({
       presets: [presetWarp({ skipResets: true })],
       safelist: classes,
     }),
+    rollupPluginReplace({
+      'process.env.NODE_ENV': JSON.stringify('production'),
+      preventAssignment: true,
+    }),
     env.mode !== 'lib' && VitEik(),
     MinifyWarpLib(),
   ],
   server: { host: '0.0.0.0', port: 3003 },
   test: {
+    globals: true,
     environment: 'happy-dom',
     include: ['./test/**'],
     exclude: ['**.json'],
+    setupFiles: ['./setup.js'],
     coverage: {
       cleanOnRerun: true,
       reporter: ['text'],
@@ -33,6 +40,7 @@ export default defineConfig((env) => ({
         'components/**/stories',
         '.minifier-plugin.js',
         'lingui.config.ts',
+        'components/**/locales',
       ],
     },
   },
@@ -51,7 +59,6 @@ function getBuildOpts(env) {
       build: {
         lib: {
           entry: {
-            tag: 'components/tag/index.js',
             tabs: 'components/tabs/index.js',
             switch: 'components/switch/index.js',
             steps: 'components/steps/index.js',